+++ lib/src/plotly.dart
@@ -32,49 +32,79 @@ class Plotly extends StatefulWidget {
     required this.viewId,
     required this.data,
     required this.layout,
-    this.displaylogo: false,
+    this.displaylogo = false,
     this.displayModeBar,
     this.editable,
     this.linkText,
-    this.responsive: true,
-    this.showLink: false,
+    this.responsive = true,
+    this.showLink = false,
     this.staticPlot,
     this.scrollZoom,
-  });
+    Key? key,
+  }) : super(key: key) {
+    divElement = DivElement()..id = viewId;
+    bodyElement.append(divElement);
+    ui.platformViewRegistry.registerViewFactory(
+        viewId,
+        (viewId) => bodyElement
+          ..style.width = '100%'
+          ..style.height = '100%');
+    plot = Plot(divElement, data, layout,
+        displaylogo: displaylogo,
+        displayModeBar: displayModeBar,
+        editable: editable,
+        linkText: linkText,
+        responsive: responsive,
+        showLink: showLink,
+        staticPlot: staticPlot,
+        scrollZoom: scrollZoom);
+  }
+
+  final bodyElement = BodyElement();
+
+  /// Where the chart will be embedded
+  late final DivElement divElement;
+  late final Plot plot;
+
+  /// Adds a new trace to an existing plot at any location in its data array.
+  void addTrace(Map trace, [int? newIndex]) => plot.addTrace(trace, newIndex);
+
+  /// Adds new traces to an existing plot at any location in its data array.
+  void addTraces(List<Map> traces, [List<int>? newIndices]) =>
+      plot.addTraces(traces, newIndices);
+
+  /// Removes a trace from a plot by specifying the index of the trace to be
+  /// removed.
+  void deleteTrace(int index) => plot.deleteTrace(index);
+
+  /// Removes traces from a plot by specifying the indices of the traces to be
+  /// removed.
+  void deleteTraces(List<int> indices) => plot.deleteTraces(indices);
+
+  /// An efficient means of changing parameters in the data array. When
+  /// restyling, you may choose to have the specified changes effect as
+  /// many traces as desired. The update is given as a single [Map] and
+  /// the traces that are effected are given as a list of traces indices.
+  /// Note, leaving the trace indices unspecified assumes that you want
+  /// to restyle *all* the traces.
+  void restyle(Map aobj, [List<int>? traces]) => plot.restyle(aobj, traces);
+
+  /// An efficient means of updating just the layout of a plot.
+  void relayout(Map aobj) => plot.relayout(aobj);
+
+  /// A method for updating both the data and layout objects at once.
+  void update(Map dataUpdate, Map layoutUpdate) =>
+      plot.update(dataUpdate, layoutUpdate);
 
   @override
   State<StatefulWidget> createState() => _PlotlyState();
 }
 
 class _PlotlyState extends State<Plotly> {
-  final bodyElement = BodyElement();
-
-  /// Where the chart will be embedded
-  late final DivElement divElement;
-
-  @override
-  void initState() {
-    super.initState();
-    divElement = DivElement()..id = widget.viewId;
-    bodyElement.append(divElement);
-  }
+  Plot? plot;
 
   @override
   Widget build(BuildContext context) {
-    ui.platformViewRegistry.registerViewFactory(
-        widget.viewId,
-        (viewId) => bodyElement
-          ..style.width = '100%'
-          ..style.height = '100%');
-    Plot(divElement, widget.data, widget.layout,
-        displaylogo: widget.displaylogo,
-        displayModeBar: widget.displayModeBar,
-        editable: widget.editable,
-        linkText: widget.linkText,
-        responsive: widget.responsive,
-        showLink: widget.showLink,
-        staticPlot: widget.staticPlot,
-        scrollZoom: widget.scrollZoom);
     return HtmlElementView(viewType: widget.viewId);
   }
 }