+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.2.0 (released 2021-12-23)
+- Improved the example significantly to show various plot interactions.  
+- Moved code from the build method to the init to improve performance.
+- Exposed most used Plotly Javascript methods directly in the Plotly Widget.  
+This makes the code cleaner.
+
 ## 0.1.1 (released 2021-09-13)
 - Fixed some to the Plotly config arguments.  Updated some warnings 
 for Flutter 2.5.  Getting some js errors in the js console.  Not sure